Firs corrections
